requirements.txt
faker==33.1.0
pg8000==1.31.5
mysql-connector-python==9.1.0
pymongo==4.10.1
neo4j==5.26.0
tqdm==4.67.1
